[refactor] 위젯 분리
